\name{neg.zero.fp}
\alias{neg.zero.fp}
\docType{data}
\title{
  Negative zero FP
}
\description{
  A data set that resulted in an error, negative FP, but actually
  numerically zero.
}
\usage{data("neg.zero.fp")}
\format{
Named list. diffs is a data table, output of aum_diffs, pred is a
numeric vector of predictions.
}
\keyword{datasets}
