\name{plot.aum_diffs}
\alias{plot.aum_diffs}
\title{plot aum diffs}
\description{Plot method for \code{\link{aum_diffs}} which shows piecewise constant error
functions. Uses \code{\link{aum_errors}} internally to compute error functions
which are plotted.}
\usage{\method{plot}{aum_diffs}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





