\name{plot.aum_line_search}
\alias{plot.aum_line_search}
\title{plot aum line search}
\description{Plot method for \code{\link{aum_line_search}} which shows AUM and threshold functions. }
\usage{\method{plot}{aum_line_search}(x, 
    ...)}
\arguments{
  \item{x}{list with class "aum_line_search".}
  \item{\dots}{ignored.}
}

\value{ggplot.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Jadon Fowler [aut] (Contributed exact line search C++ code)}





