\name{plot.aum_line_search_grid}
\alias{plot.aum_line_search_grid}
\title{plot aum line search grid}
\description{Plot method for \code{\link{aum_line_search_grid}} which shows AUM and threshold
functions, along with grid points for checking.}
\usage{\method{plot}{aum_line_search_grid}(x, 
    ...)}
\arguments{
  \item{x}{list with class "aum_line_search_grid".}
  \item{\dots}{ignored.}
}

\value{ggplot.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Jadon Fowler [aut] (Contributed exact line search C++ code)}





