\name{set_loss_plot}
\alias{set_loss_plot}
\title{set loss plot}
\description{plot subtrain/validation loss.}
\usage{set_loss_plot(loss.dt, 
    set.colors = c(subtrain = "black", 
        validation = "red"))}
\arguments{
  \item{loss.dt}{loss.dt }
  \item{set.colors}{set.colors }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Jadon Fowler [aut] (Contributed exact line search C++ code)}





